backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
Pillow==11.1.0
mistune==3.0.2
python-multipart==0.0.20
aiofiles==24.1.0
